requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=1.0.0
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2
pandas>=2.0.0
pdfplumber>=0.10.3
matplotlib>=3.7.1
